refactor(cli): sweep cross-group shared helpers into core.py (#495) (#530)
* refactor(cli): sweep cross-group shared helpers into core.py (#495)

Phase 0.5 of #495. Relocate cross-group private helpers from __main__.py
into core.py so the parallel Phase 1 domain extractions don't collide on
them. Pure relocation, CLI surface byte-identical, all behavior unchanged.

Moved (def relocated to core.py, deleted from __main__.py, re-imported via
the existing `from .core import (...)` block):

- _run_remote ............ session + panes + diagnostics (29 call sites)
- _default_portal_url .... services + session + panes + diagnostics
- _get_portal_url ........ services (say) + system (open, notify)
- _check_portal_health ... transitive dep of _get_portal_url (portal-client)
- _portal_auth_headers ... transitive dep of _notify_* / _post_desktop_*
- _notify_portal_sessions_changed . session (new) + panes (kill) + diagnostics
- _post_desktop_notification ...... services (say) + system (notify-user)
- _resolve_session_type_from_args . session (new) + panes (spawn)
- KIND_DEFAULT_POSTURE ... constant used by the above + cmd_new
- _add_posture_harness_flags ...... build_parser INFRA (new/worktree/spawn)
- _git_behind_origin ..... diagnostics (doctor) + system (rebuild)
- _get_agentwire_path .... panes (spawn) + services/system (portal hooks)

Built by dotdev.dev

* refactor(cli): move _start_portal_local + _install_global_tmux_hooks to core (#495)

Per review of #530: _start_portal_local is genuinely cross-group
(cmd_portal_start=services, cmd_up=system). It and its sole-caller
transitive dep _install_global_tmux_hooks depend only on _get_agentwire_path
(already in core in this PR) — no cycle, no hooks-installer drag.

Built by dotdev.dev
